requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# LLM Integration
openai==1.3.0
anthropic==0.7.0
langchain==0.1.0
langchain-openai==0.0.2

# Agent Frameworks
autogen==0.2.0

# System Operations
psutil==5.9.6
pyautogui==0.9.54
pygetwindow==0.0.9

# Web Scraping
beautifulsoup4==4.12.2
scrapy==2.11.0
playwright==1.40.0
requests==2.31.0

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Utilities
loguru==0.7.2
tenacity==8.2.3
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
pylint==3.0.3
mypy==1.7.1

# Monitoring & Logging
prometheus-client==0.19.0
